move search term criteria matching to service
